backend/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
django-cors-headers==4.7.0
whitenoise==6.8.2
dj-database-url==2.3.0
bcrypt==4.3.0
boto3==1.38.32
pikepdf==9.7.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
sentry-sdk[django]>=2.0.0
django-redis>=5.4.0
celery>=5.4.0
celery[redis]>=5.4.0
django-celery-beat>=2.6.0
python-magic>=0.4.27
pyclamd>=0.4.0
openpyxl>=3.1.0
